backend/requirements.txt
clickhouse_connect==0.6.17
concurrent_log_handler==0.9.24
datasketch==1.6.4
dnspython==2.4.2
elasticsearch==7.10.1
email-validator==2.1.0.post1
fire==0.5.0
Flask==2.2.2
Flask_Cors==4.0.0
Flask_JWT_Extended==4.5.3
flask_mongoengine==1.0.0
flask-wtf==1.2.1
gunicorn
joblib==1.3.2
langchain==0.2.0
loguru==0.7.2
marshmallow==3.20.1
mongoengine==0.27.0
numpy==1.26.4
openai==0.28.1
opencv-python-headless
pandas==2.2.2
pdf2docx==0.5.6
pdfminer
pdfplumber==0.11.0
peewee==3.17.5
prometheus_flask_exporter==0.23.0
pymongo==4.5.0
pymupdf==1.24.4
portalocker==2.8.2
pymupdfb==1.24.3
pypandoc==1.12
pypdfium2==4.30.0
python_docx==1.0.1
Requests==2.32.2
rouge_chinese==1.0.3
scikit_learn==1.3.2
setproctitle==1.3.3
termcolor==2.3.0
threadpoolctl==3.2.0
tqdm==4.66.4
transformers==4.41.0
webargs==8.3.0
Werkzeug==2.2.2
wtforms==3.1.0
nacos-sdk-python==1.0.0
oss2==2.18.6
tldextract